+++ types/next-auth.d.ts
@@ -6,6 +6,7 @@ declare module 'next-auth' {
       id: string;
       name?: string | null;
       email?: string | null;
+      profile?: any;
       image?: string | null;
     };
     accessToken?: string;
